Updated data schema
Updated the data schema to use objects instead of arrays. Moved all availability helper functions into utils/availability.js and out of schedule.js
