+++ src/savedAvailability.json
@@ -1,3 +1,1 @@
-{
-  "data": []
-}
+{}
