+++ src/utils/schedule.js
@@ -76,44 +76,6 @@ const createSchedule = (daysList, time) => {
 };
 
 /**
- * Create Unavailable object
- * @param {Dayjs} start
- * @param {Dayjs} end
- * @param {string} reason
- * @returns {Object}
- */
-const createUnavailability = (start, end, reason) => {
-    if(dayjs(start).isAfter(dayjs(end)))
-        throw new ScheduleError('End Date/Time must be after Start Date/Time');
-
-    return {
-            from: start,
-            to: end,
-            reason: reason
-        };
-};
-
-/**
- * Create Available object
- * @param {Dayjs} start
- * @param {Dayjs} end
- * @param {string} days
- * @returns {Object}
- */
-const createAvailability = (start, end, days) => {
-    if(dayjs(start).isAfter(dayjs(end)))
-        throw new ScheduleError('End Date/Time must be after Start Date/Time');
-
-    if(!days)
-        days = 'Monday-Friday';
-
-    return {from: start,
-            to: end,
-            days: days
-        };
-};
-
-/**
  * Remove any duplicate time entries
  * @param {*} schedules 
  * @returns
